+++ radio/single_radio_pin_mapping_odroid_m1_default.cc
@@ -3,8 +3,8 @@
 namespace tvsc::radio {
 
 const char* SingleRadioPinMapping::board_name() { return "RF69HCW 433 Odroid M1"; }
-int SingleRadioPinMapping::reset_pin() { return 106; }
-int SingleRadioPinMapping::chip_select_pin() { return 14; }
-int SingleRadioPinMapping::interrupt_pin() { return 12; }
+uint8_t SingleRadioPinMapping::reset_pin() { return 106; }
+uint8_t SingleRadioPinMapping::chip_select_pin() { return 14; }
+uint8_t SingleRadioPinMapping::interrupt_pin() { return 12; }
 
 }  // namespace tvsc::radio
