+++ radio/single_radio_pin_mapping_teensy41_default.cc
@@ -3,8 +3,8 @@
 namespace tvsc::radio {
 
 const char* SingleRadioPinMapping::board_name() { return "RF69HCW 433 Teensy41"; }
-int SingleRadioPinMapping::reset_pin() { return 9; }
-int SingleRadioPinMapping::chip_select_pin() { return 10; }
-int SingleRadioPinMapping::interrupt_pin() { return 24; }
+uint8_t SingleRadioPinMapping::reset_pin() { return 9; }
+uint8_t SingleRadioPinMapping::chip_select_pin() { return 10; }
+uint8_t SingleRadioPinMapping::interrupt_pin() { return 24; }
 
 }  // namespace tvsc::radio
